requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Werkzeug>=2.3.0
SQLAlchemy>=2.0.0
reportlab>=4.0.0
openpyxl>=3.1.0
Pillow>=10.0.0
python-dateutil>=2.8.0
bcrypt>=4.0.0
email-validator>=2.0.0
Jinja2>=3.1.0
opencv-contrib-python-headless>=4.8.0
numpy>=1.24.0
gunicorn>=21.2.0

psycopg2-binary>=2.9.0
